fix: dist not generating full models
pesky bug that is caused because some time ago we added the "browser" property to the package.json this is helpful for webpack and other build systems to identify the browser version of the library to use depending on the target. Unfortunately if there is a "browser" property browserify just looks at that instead of our src/index.js. Its a bit of a catch-22 and there might be a better solution that I'll have to look into more later