:boom: revert filters to work how they did previously
Found out that the filter types weren't being enforced unless inlined. Essentially rendering the feature useless.

This is because object literals undergo a special treatement called "excess property checks", if a property that doesn't exist is added it throws an error. This is what we want, but only happens if you inline your object.